Move UI components into their own classes
